refactor(lib): better header integration
