+++ lib/my/compile_flags.txt
@@ -1,2 +1,2 @@
 -I
-include/
+.
