+++ lib/my/sort/my_sort_int_array.c
@@ -5,7 +5,7 @@
 ** Sorts an integer array in ascending order
 */
 
-#include "my_sort.h"
+#include "sort/my_sort.h"
 
 static void swap_if_greater(int *array, int j)
 {
