+++ lib/my/stdio/my_stdio.h
