+++ lib/my/strings/my_str_isalpha.c
@@ -5,8 +5,8 @@
 ** Returns if a string only contains alpha char or not
 */
 
-#include "my_strings.h"
-#include "my_ctype.h"
+#include "strings/my_strings.h"
+#include "ctype/my_ctype.h"
 #include <stddef.h>
 
 int my_str_isalpha(char const *str)
