+++ lib/my/strings/my_strings.h
